+++ src/UserVideoAbstraction.h
@@ -1,3 +1,11 @@
+//****************************************
+//user: PianoCoder
+//Create date: 2015/01/01
+//Class name:  UserVideoAbstraction(...) class implemention
+//Discription: implement the interface for all the video abstraction operation
+//Update:      2015/01/17
+//****************************************
+
 #pragma once
 #include <iostream>
 #include <string>
@@ -22,7 +30,7 @@ public:
 	__declspec(dllexport) void UsersetGpu(bool isgpu);													// 设置是否使用GPU
 	__declspec(dllexport) void UsersetROI(Rect rectroi);											// 设置是否选择感兴趣的区域
 	__declspec(dllexport) void UsersetIndex(bool isindex);
-	__declspec(dllexport) void UsersetSingleMinArea(int area);
-	__declspec(dllexport) void UsersetMinArea(int area);
+	__declspec(dllexport) void UsersetSingleMinArea(float area);
+	__declspec(dllexport) void UsersetMinArea(float area);
 	__declspec(dllexport) void UserGetKeyFrame(string filepath, int frameCount);
 };
